backend/requirements.txt
flask
flask-cors
httpx
python-dotenv
PyPDF2  
tqdm
google-generativeai